+++ src/dsgvo/dsgvo_mail.php
@@ -1,5 +1,6 @@
 <?php include dirname(__DIR__) . '/header.php';?>
 <?php require dirname(__DIR__) . "/misc/helpcenter.php"; ?>
+<?php require_permission("READ","DSGVO","EMAIL_TEMPLATES") ?>
 <?php
 if(empty($_GET['n']) || !in_array($_GET['n'], $available_companies)){ //eventually STRIKE
     $conn->query("UPDATE userdata SET strikeCount = strikecount + 1 WHERE id = $userID");
@@ -9,7 +10,9 @@ if(empty($_GET['n']) || !in_array($_GET['n'], $available_companies)){ //eventual
 }?>
 <div class="page-header-fixed">
 <div class="page-header"><h4><?php echo $lang["EMAIL_TEMPLATES"] ?>
+<?php if(has_permission("WRITE","DSGVO","EMAIL_TEMPLATES")): ?>
  <div class="page-header-button-group"><form method="POST" style="display:inline"><button type="submit" name="create_report" class="btn btn-default"><i class="fa fa-plus"></i></button></form>
+<?php endif ?>
 </h4></div>
 </div>
 <div class="page-content-fixed-130">
@@ -20,34 +23,36 @@ $cmpID = intval($_GET['n']);
 $action = $action_id = '';
 if(isset($_POST['create_report'])){ $action = 'new'; }
 if(isset($_POST['edit_report'])){ $action = 'edit'; $action_id = intval($_POST['edit_report']); }
+if(has_permission("WRITE","DSGVO","EMAIL_TEMPLATES")){
+    if($action && !empty($_POST['report_content']) && !empty($_POST['report_name'])){
+        $templateName = test_input($_POST['report_name']);
+        $templateContent = $_POST['report_content'];
+        if($action == 'new'){
+            $stmt = $conn->prepare("INSERT INTO templateData (name, htmlCode, type, userIDs) VALUES(?, ?, 'document', $cmpID)");
+        } else {
+            $stmt = $conn->prepare("UPDATE templateData SET name = ?, htmlCode = ? WHERE id = $action_id");
+        }
+        $stmt->bind_param("ss", $templateName, $templateContent);
+        $stmt->execute();
 
-if($action && !empty($_POST['report_content']) && !empty($_POST['report_name'])){
-    $templateName = test_input($_POST['report_name']);
-    $templateContent = $_POST['report_content'];
-    if($action == 'new'){
-        $stmt = $conn->prepare("INSERT INTO templateData (name, htmlCode, type, userIDs) VALUES(?, ?, 'document', $cmpID)");
-    } else {
-        $stmt = $conn->prepare("UPDATE templateData SET name = ?, htmlCode = ? WHERE id = $action_id");
-    }
-    $stmt->bind_param("ss", $templateName, $templateContent);
-    $stmt->execute();
+        $action = '';
 
-    $action = '';
-
-    if($conn->error){
-        echo '<div class="alert alter-danger"><a href="" data-dismiss="alert" class="close">&times;</a></div>';
-    } else {
-        echo '<div class="alert alert-success"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$lang['OK_SAVE'].'</div>';
-    }
-} elseif(!empty($_POST['delete_report'])){
-    $val = intval($_POST['delete_report']);
-    $conn->query("DELETE FROM templateData WHERE id = $val");
-    if($conn->error){
-        echo '<div class="alert alert-danger"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$conn->error.'</div>';
-    } else {
-        echo '<div class="alert alert-success"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$lang['OK_ADD'].'</div>';
+        if($conn->error){
+            echo '<div class="alert alter-danger"><a href="" data-dismiss="alert" class="close">&times;</a></div>';
+        } else {
+            echo '<div class="alert alert-success"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$lang['OK_SAVE'].'</div>';
+        }
+    } elseif(!empty($_POST['delete_report'])){
+        $val = intval($_POST['delete_report']);
+        $conn->query("DELETE FROM templateData WHERE id = $val");
+        if($conn->error){
+            echo '<div class="alert alert-danger"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$conn->error.'</div>';
+        } else {
+            echo '<div class="alert alert-success"><a href="#" data-dismiss="alert" class="close">&times;</a>'.$lang['OK_ADD'].'</div>';
+        }
     }
 }
+
 ?>
 
 <form method="POST">
@@ -63,7 +68,9 @@ if($action && !empty($_POST['report_content']) && !empty($_POST['report_name']))
         echo '<tr>';
         echo '<td>'.$row['name'].'</td>';
         echo '<td>';
-        echo '<button type="submit" name="delete_report" value="'.$row['id'].'" class="btn btn-default" title="'.$lang['DELETE'].'"><i class="fa fa-trash-o"></i></button> ';
+        if(has_permission("WRITE","DSGVO","EMAIL_TEMPLATES")){
+            echo '<button type="submit" name="delete_report" value="'.$row['id'].'" class="btn btn-default" title="'.$lang['DELETE'].'"><i class="fa fa-trash-o"></i></button> ';
+        }
         echo '<button type="submit" name="edit_report" value="'.$row['id'].'" class="btn btn-default" title="'.$lang['EDIT'].'" ><i class="fa fa-pencil"></i></button>';
         echo '</td>';
         echo '</tr>';
@@ -90,7 +97,7 @@ if($action == 'edit'){
 ?>
 <div class="row">
     <div class="col-xs-6"><label>Name</label><input type="text" class="form-control" placeholder="Name of Template (Required)" name="report_name" value="<?php echo $templateName; ?>" /></div>
-    <div class="col-xs-3"><label><?php echo $lang["SAVE"] ?></label><br><button type="submit" class="btn btn-warning"><i class="fa fa-floppy-o"></i></button></div>
+    <div class="col-xs-3"><?php if(has_permission("WRITE","DSGVO","EMAIL_TEMPLATES")): ?><label><?php echo $lang["SAVE"] ?></label><br><button type="submit" class="btn btn-warning"><i class="fa fa-floppy-o"></i></button><?php endif ?></div>
 </div>
 <div class="row">
     <div class="col-sm-9" style="max-width:790px;"><textarea name="report_content"><?php echo $templateContent; ?></textarea></div>
